% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifyHelpers.R
\name{add}
\alias{add}
\title{Create function that adds text elements to vector}
\usage{
add(..., .pos = "bottom")
}
\arguments{
\item{...}{Elements to add.}

\item{.pos}{Either \"top\" or \"bottom\". Decides if new text is prepended or appended to existing text.}
}
\value{
A function that adds the specified text to character vectors
}
\description{
Create function that adds text elements to vector
}
\examples{
myfun <- add("b","d")
myfun("a")
myfun2 <- add("b","d",.pos="top")
myfun2("a")
}
