% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addBlocks.R
\name{addBlocks}
\alias{addBlocks}
\title{Add blocking info to parameter set}
\usage{
addBlocks(pars, col.model = "model")
}
\arguments{
\item{pars}{The parameter, as returned by `NMreadExt()`}

\item{col.model}{Name of the model name column.}
}
\description{
Add blocking info to parameter set
}
\keyword{internal}
