% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addClass.R
\name{addClass}
\alias{addClass}
\title{Add class if not already present}
\usage{
addClass(data, class)
}
\arguments{
\item{data}{The object to add class to}

\item{class}{The class to add (character)}
}
\value{
Object with additional class
}
\description{
Add class if not already present
}
\keyword{internal}
