% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addEVID2.R
\name{addEVID2}
\alias{addEVID2}
\title{Add simulation records to dosing records}
\usage{
addEVID2(doses, time.sim, CMT)
}
\arguments{
\item{doses}{dosing records Nonmem style}

\item{time.sim}{A numerical vector with simulation times}

\item{CMT}{The compartment in which to insert the EVID=2 records}
}
\description{
Add simulation records to dosing records
}
\examples{
(doses1 <- NMcreateDoses(TIME=c(0,12,24,36),AMT=c(2,1)))
addEVID2(doses1,time.sim=seq(0,28,by=4),CMT=2)
}
