% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMctl.R
\name{as.NMctl}
\alias{as.NMctl}
\title{Convert object to class NMctl}
\usage{
as.NMctl(x, ...)
}
\arguments{
\item{x}{object to convert}

\item{...}{Not used}
}
\value{
An object of class `NMctl`.
}
\description{
Convert object to class NMctl
}
