% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/needRun.R
\name{callArgs}
\alias{callArgs}
\title{get arguments passed to a function}
\usage{
callArgs(which = -1)
}
\arguments{
\item{which}{The number of environment levels to move. Default is
1 because that means the result is concerns the
function/environment in which callArgs() is executed.}
}
\value{
List of arguments
}
\description{
get arguments passed to a function
}
