% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMexecDirectory.R
\name{callNonmemDirect}
\alias{callNonmemDirect}
\title{Generate system command to call Nonmem directly}
\usage{
callNonmemDirect(file.mod, path.nonmem)
}
\description{
Generate system command to call Nonmem directly
}
\keyword{internal}
