% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkTimes.R
\name{checkTimes}
\alias{checkTimes}
\title{Test if file modification times indicate that Nonmem models should
be re-run}
\usage{
checkTimes(
  file,
  use.input = TRUE,
  nminfo.input = NULL,
  file.mod,
  tz.lst = NULL,
  use.tmp = TRUE
)
}
\arguments{
\item{file}{Path to Nonmem-created file. Typically an output
control stream.}

\item{use.input}{Scan input data for updates too? Default is TRUE.}

\item{nminfo.input}{If you do want to take into account input data
but avoid re-reading the information, you can pass the NMdata
meta data object.}

\item{file.mod}{The input control stream}

\item{tz.lst}{If files are moved around on or between file
systems, the file modification time may not be reflective of
the Nonmem runtime. In that case, you can choose to extract
the time stamp from the output control stream. The issue is
that Nonmem does not write the time zone, so you have to pass
that to checkTimes if this is wanted.}
}
\description{
Test if file modification times indicate that Nonmem models should
be re-run
}
\keyword{internal}
