% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMsim_helpers.R
\name{cleanStrings}
\alias{cleanStrings}
\title{Drop spaces and odd characters}
\usage{
cleanStrings(x)
}
\arguments{
\item{x}{a string to clean}
}
\description{
Drop spaces and odd characters
}
\keyword{internal}
