% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMreadInits.R
\name{count_ij}
\alias{count_ij}
\title{Assign i and j indexes based on parameter section text}
\usage{
count_ij(res)
}
\arguments{
\item{res}{elements as detected by `NMreadInits()`}
}
\description{
Assign i and j indexes based on parameter section text
}
\keyword{internal}
