% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyclicDoses.R
\name{cyclicDoses}
\alias{cyclicDoses}
\title{create intermittent dosing regimen}
\usage{
cyclicDoses(hours.of.day, cycles.days, days.total, dose)
}
\arguments{
\item{hours.of.day}{dosing time of day (hours)}

\item{cycles.days}{list of numeric vectors. The vectors represent
nested regimens. See details.}

\item{days.total}{Total number of days for the regimen to
continue.}

\item{dose}{Will be copied to amt column}
}
\description{
create intermittent dosing regimen
}
\details{
\itemize{ \item list(c(4,3)) means three days on, four
    days off.  \item list(c(4,3),c(3,1)) means three days on, four
    days off for three cycles, then one cycle (of 4+3 days) off.
    }
}
\examples{
Complicated like ph2
dt <- cyclicDoseTab(c(0,12),cycles.days=list(c(4,3),c(21,7)),days.total=56,dose=400)
ggplot(dt,aes(times/24,amt))+geom_point()+geom_vline(xintercept=c(0,28,56))
}
