% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcastSe.R
\name{dcastSe}
\alias{dcastSe}
\title{A standard-evaluation interface to `data.table::dcast()`}
\usage{
dcastSe(data, l, r, ...)
}
\arguments{
\item{data}{data set to transpose (widen)}

\item{l}{left-hand side variables as character vector. Result will
be long/vertical in these variables.}

\item{r}{left-hand side variables as character vector. Result will
be wide in these variables.}

\item{...}{Additional arguments paseed to `data.table::dcast()`.}
}
\description{
A standard-evaluation interface to `data.table::dcast()`
}
\keyword{internal}
