% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/needRun.R
\name{digestElements}
\alias{digestElements}
\title{Get digest checksum of object elements}
\usage{
digestElements(obj, funs)
}
\arguments{
\item{obj}{An object with elements to run digest on.}

\item{funs}{Named list of functions to be applied to elements (matched on names) in `obj`. Optional.}
}
\value{
A data.table with hashes/checksums.
}
\description{
Get digest checksum of object elements
}
