% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandCovs.R
\name{expandCovs}
\alias{expandCovs}
\title{Create data set where each covariate is univariately varied (see `forestDefineCovs()`)}
\usage{
expandCovs(...)
}
\arguments{
\item{...}{Passed to `forestDefineCovs()`}
}
\value{
A data.frame
}
\description{
Create data set where each covariate is univariately varied (see `forestDefineCovs()`)
}
