% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genPhiFile.R
\name{genPhiFile}
\alias{genPhiFile}
\title{Generate a .phi file for further simulation with Nonmem}
\usage{
genPhiFile(data, file)
}
\arguments{
\item{data}{A dataset that contains "ID" and all ETAs. This can be
obtained by `NMdata::NMscanData`.}

\item{file}{Path to the .phi file to be written.}
}
\description{
This will typically be used in a couple of different
situations. One is if a number of new subjects have been simulated
and their ETAs should be reused in subsequent simulations. Another
is internally by NMsim when simulating new subjects from models
estimated with SAEM.
}
