% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputArchive.R
\name{inputArchiveDefault}
\alias{inputArchiveDefault}
\title{Default location of input archive file}
\usage{
inputArchiveDefault(file)
}
\arguments{
\item{file}{Path to input or output control stream.}
}
\value{
A file name (character)
}
\description{
Default location of input archive file
}
