% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMsimTab_utils.R
\name{is.NMsimTab}
\alias{is.NMsimTab}
\title{Check if an object is 'NMsimTab'}
\usage{
is.NMsimTab(x)
}
\arguments{
\item{x}{Any object}
}
\value{
logical if x is an 'NMsimTab' object
}
\description{
Check if an object is 'NMsimTab'
}
