% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMreadInits.R
\name{itriag}
\alias{itriag}
\title{Row numbers of elements in a triangular representation of a symmetric matrix}
\usage{
itriag(blocksize, istart = 1, diag = "lower")
}
\description{
Row numbers of elements in a triangular representation of a symmetric matrix
}
\keyword{internal}
