% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMreadInits.R
\name{jtriag}
\alias{jtriag}
\title{Column numbers of elements in a triangular representation of a symmetric matrix}
\usage{
jtriag(blocksize, istart = 1, diag = "lower")
}
\description{
Column numbers of elements in a triangular representation of a symmetric matrix
}
\keyword{internal}
