% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifyModel.R
\name{modifyModel}
\alias{modifyModel}
\title{Internal method for handling modify.model argument to NMsim}
\usage{
modifyModel(modify.model, dt.models)
}
\arguments{
\item{modify.model}{A list}

\item{dt.models}{a data.table}
}
\description{
Internal method for handling modify.model argument to NMsim
}
\keyword{internal}
