% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/needRun.R
\name{needRun}
\alias{needRun}
\title{Determine whether re-run of a function is necessary}
\usage{
needRun(path.res, path.digest, funs, which = -2)
}
\arguments{
\item{path.res}{Path to function results output file}

\item{path.digest}{Path to the file containing digests to compare
to.}

\item{funs}{Named list of functions to apply to arguments}

\item{which}{Number of environment levels to jump to evaluate the
arguments}
}
\description{
Compares arguments against stored checksums or output of functions
of them. If changes found, the function should be re-run.
}
