% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMexecDirectory.R
\name{NMexecDirectory}
\alias{NMexecDirectory}
\title{Execute Nonmem inside a dedicated directory}
\usage{
NMexecDirectory(file.mod, path.nonmem, files.needed, dir.data = "..")
}
\arguments{
\item{dir.data}{If NULL, data will be copied into the temporary
directory, and Nonmem will read it from there. If not NULL,
dir.data must be the relative path from where Nonmem is run to
where the input data file is stored. This would be ".." if the
run directory is created in a directory where the data is
stored.}
}
\value{
A bash shell script for execution of Nonmem
}
\description{
Execute Nonmem inside a dedicated directory
}
\keyword{internal}
