% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMexecSingle.R
\name{NMexecSingle}
\alias{NMexecSingle}
\title{execute nonmem while also archiving input data}
\usage{
NMexecSingle(
  file.mod,
  sge = TRUE,
  file.data.archive,
  nc = 64,
  dir.data = NULL,
  wait = FALSE,
  args.execute
)
}
\arguments{
\item{sge}{Use the sge queing system. Default is TRUE. Disable for
quick models not to wait.}

\item{file.data.archive}{A function of the model file path to
generate the path in which to archive the input data as
RDS. Set to NULL not to archive the data.}
}
\description{
execute nonmem while also archiving input data
}
\details{
Use this to read the archived input data when retrieving
    the nonmem results
    NMdataConf(file.data=function(x)fnExtension(fnAppend(x,"input"),".rds"))
}
