% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMgenPNM.R
\name{NMgenPNM}
\alias{NMgenPNM}
\title{Generate PNM file for sge clusters}
\usage{
NMgenPNM(nc, file)
}
\arguments{
\item{nc}{number of cores wanted}

\item{file}{The file path to write the result to}
}
\value{
The file path (character string)
}
\description{
Generate PNM file for sge clusters
}
\keyword{internal}
