% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMreadExt.R
\name{NMreadExt}
\alias{NMreadExt}
\title{Read information from Nonmem ext files}
\usage{
NMreadExt(path.ext)
}
\arguments{
\item{path.ext}{Path to the ext file}
}
\value{
A list with a final parameter table and a table of the iterations
}
\description{
Read information from Nonmem ext files
}
\keyword{internal}
