% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMreadFilters.R
\name{NMreadFilters}
\alias{NMreadFilters}
\title{Read data filters from a NONMEM model}
\usage{
NMreadFilters(file, lines, filters.only = TRUE, as.fun)
}
\arguments{
\item{file}{Control stream path}

\item{lines}{Control stream lines if already read from file}

\item{filters.only}{Return the filters only or also return the remaining text in a separate object? If `FALSE`, a list with the two objects is returned.}

\item{as.fun}{Function to run on the tables with filters.}
}
\description{
Read data filters from a NONMEM model
}
\keyword{internal}
