% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMreadPhi.R
\name{NMreadPhi}
\alias{NMreadPhi}
\title{Read information from Nonmem ext files}
\usage{
NMreadPhi(file)
}
\arguments{
\item{file}{Path to the phi file}
}
\value{
A list with a final parameter table and a table of the iterations
}
\description{
Read information from Nonmem ext files
}
\keyword{internal}
