% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMreadSimModTab.R
\name{NMreadSimModTabOne}
\alias{NMreadSimModTabOne}
\title{Read simulation results from an rds or a NMsimModTab object}
\usage{
NMreadSimModTabOne(
  modtab,
  check.time = FALSE,
  dir.sims,
  wait = FALSE,
  quiet = FALSE,
  skip.missing = FALSE,
  progress,
  read.fst = NULL,
  fast.tables = NULL,
  carry.out = NULL,
  as.fun
)
}
\arguments{
\item{check.time}{If found, check whether `fst` file modification time is
newer than `rds` file. The `fst` is generated based on information in
`rds`, but notice that some systems don't preserve the file modification
times. Becasue of that, `check.time` is `FALSE` by default.}

\item{dir.sims}{By default, `NMreadSim` will use information about the
relative path from the results table file (`_MetaData.rds`) to the Nonmem
simulation results. If these paths have changed, or for other reasons this
doesn't work, you can use the `dir.sims` argument to specify where to find
the Nonmem simulation results. If an `.fst` file was already generated and
is found next to the `_MetaData.rds`, the path to the Nonmem simulation
results is not used.}

\item{wait}{If simulations seem to not be done yet, wait for them to finish?
If not, an error will be thrown. If you choose to wait, the risk is
results never come. `NMreadSim` will be waiting for an `lst` file. If
Nonmem fails, it will normally generate an `lst` file. But if `NMTRAN`
fails (checks of control stream prior to running Nonmem), the `lst` file
is not generated. Default is not to wait.}

\item{quiet}{Turn off some messages about what is going on? Default is to
report the messages.}

\item{skip.missing}{Skip models where results are not available? Default is
`FALSE` meaning an error will be thrown if one or more models do not have
completed results.}

\item{progress}{Track progress? Default is `TRUE` if `quiet` is FALSE and
more than one model is being read. The progress tracking is based on the
number of models completed/read, not the status of the individual models.}

\item{carry.out}{Variables from input data that should be included in
results. This feature is only available if `NMsim()` was called with the
`table.vars` argument. Also, it will only be used when reading results
from output tables. To save time and disk space, `NMreadSim()` will store
compressed `.fst` files, and when these are subsequently read by
`NMreadSim()` `carry.out` is not used. To re-read the output tables and
regenerate the `.fst` file, use `reread.tmp=TRUE`.}

\item{as.fun}{The default is to return data as a data.frame. Pass a function
(say `tibble::as_tibble`) in as.fun to convert to something else. If
data.tables are wanted, use as.fun="data.table". The default can be
configured using NMdataConf.}
}
\description{
Read simulation results from an rds or a NMsimModTab object
}
\keyword{internal}
