% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMreadSimOne.R
\name{NMreadSimRes}
\alias{NMreadSimRes}
\title{Read simulation results from data.frames or fst files}
\usage{
NMreadSimRes(x)
}
\arguments{
\item{x}{a data set or a fst file}
}
\description{
Read simulation results from data.frames or fst files
}
\keyword{internal}
