% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMreadSimOne.R
\name{NMreadSimResOne}
\alias{NMreadSimResOne}
\title{read one sim element. This will be run in lapply in NMreadSim.}
\usage{
NMreadSimResOne(x)
}
\arguments{
\item{x}{A path to an fst file or a data set}
}
\value{
A data.table
}
\description{
read one sim element. This will be run in lapply in NMreadSim.
}
\keyword{internal}
