% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMreadSizes.R
\name{NMreadSizes}
\alias{NMreadSizes}
\title{Read SIZES info from a control stream}
\usage{
NMreadSizes(file.mod = NULL, lines = NULL)
}
\arguments{
\item{file.mod}{Control stream path.}

\item{lines}{Character vector with control stream file.}
}
\value{
A list with SIZES parameter values
}
\description{
Read SIZES info from a control stream
}
\keyword{internal}
