% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMreplaceInits.R
\name{NMreplaceInits}
\alias{NMreplaceInits}
\title{Replace initial values in Nonmem control stream}
\usage{
NMreplaceInits(inits, fix = FALSE, ...)
}
\arguments{
\item{inits}{A data.frame with new initial estimates, same style
as returned by NMdata::NMreadExt. Column` par.type` can contain
elements THETA, OMEGA, SIGMA.}

\item{fix}{Fix the initial values? Default is not to.}

\item{...}{Passed to NMdata::NMwriteSection. This is important for
NMreplaceInits to run at all.}
}
\value{
The modified control stream
}
\description{
Replace initial values in Nonmem control stream
}
\keyword{internal}
