% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMrunLin.R
\name{NMrunLin}
\alias{NMrunLin}
\title{Internal function to run Nonmem on linux}
\usage{
NMrunLin(
  fn.mod,
  dir.mod.abs,
  exts.cp,
  meta.tables,
  path.nonmem,
  clean,
  sge,
  nc,
  pnm,
  fun.post = NULL
)
}
\arguments{
\item{fn.mod}{Just the file name, not including path}
}
\description{
Internal function to run Nonmem on linux
}
\keyword{internal}
