% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMseed.R
\name{NMseed}
\alias{NMseed}
\title{Add seed string to simulation model data.table}
\usage{
NMseed(models, nseeds, dist, values, fun.seed = seedFunDefault)
}
\arguments{
\item{models}{A data.frame containing model paths etc as created
by \code{NMsim()}.}

\item{nseeds}{Number of seeds in each simulation control
stream. Default is to match length of dist.}

\item{dist}{Distribution of random sources. These character
strings will be pasted directly into the Nonem control streams
after the seed values. Default is "" which means one normal
distribution. \code{dist=c("","UNIFORM")} will give two seeds
with random sources following a normal and a uniform
distribution.}

\item{values}{Optionally, seed values. This can be a data.frame
with as many columns as random sources.}
}
\value{
An updated data.table with simulation model information
    including seed strings.
}
\description{
This is an internal NMsim function.
}
\keyword{internal}
