% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMsim_asis.R
\name{NMsim_asis}
\alias{NMsim_asis}
\title{Simulation method that uses the provided control stream as is}
\usage{
NMsim_asis(file.sim, file.mod, data.sim)
}
\arguments{
\item{file.sim}{See \code{?NMsim}.}

\item{file.mod}{See \code{?NMsim}.}

\item{data.sim}{See \code{?NMsim}.}
}
\value{
Path to simulation control stream
}
\description{
The simplest of all method. It does nothing (but again,
\code{NMsim} handles `$INPUT`, `$DATA`, `$TABLE` and more. Use
this for instance if you already created a simulation (or
estimation actually) control stream and want NMsim to run it on
different data sets.
}
