% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMsim_known.R
\name{NMsim_known}
\alias{NMsim_known}
\title{NMsim_known is an old name for NMsim_EBE()}
\usage{
NMsim_known(...)
}
\arguments{
\item{...}{Everything passed to NMsim_EBE()}
}
\description{
NMsim_known is an old name for NMsim_EBE()
}
\keyword{internal}
