% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMsim_typical.R
\name{NMsim_typical}
\alias{NMsim_typical}
\title{Typical subject simiulation method}
\usage{
NMsim_typical(file.sim, file.mod, data.sim, return.text = FALSE)
}
\arguments{
\item{file.sim}{See \code{?NMsim}.}

\item{file.mod}{See \code{?NMsim}.}

\item{data.sim}{See \code{?NMsim}.}

\item{return.text}{If TRUE, just the text will be returned, and
resulting control stream is not written to file.}
}
\value{
Path to simulation control stream
}
\description{
Like \code{NMsim_default} but with all ETAs=0, giving a
"typical subject" simulation. Do not confuse this with a
"reference subject" simulation which has to do with covariate
values. Technically all ETAs=0 is obtained by replacing
\code{$OMEGA} by a zero matrix.
}
\keyword{internal}
