% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMsimCheckMod.R
\name{NMsimCheckMod}
\alias{NMsimCheckMod}
\title{Check a simulation control streams for things that can cause
trouble in NMsim}
\usage{
NMsimCheckMod(file.mod, lines)
}
\arguments{
\item{file.mod}{A control stream to check}

\item{lines}{The control stream as text lines. Only use of of
`file.mod` and `lines`.}
}
\description{
Check a simulation control streams for things that can cause
trouble in NMsim
}
\keyword{Don't}
\keyword{export}
\keyword{internal}
