% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMsimModels_utils.R
\name{unNMsimModels}
\alias{unNMsimModels}
\alias{is.NMsimModels}
\alias{NMsimModelsOperations}
\alias{merge.NMsimModels}
\alias{t.NMsimModels}
\alias{dimnames.NMsimModels}
\alias{rbind.NMsimModels}
\alias{cbind.NMsimModels}
\title{Remove NMsimModels class and discard NMsimModels meta data}
\usage{
unNMsimModels(x)

is.NMsimModels(x)

\method{merge}{NMsimModels}(x, ...)

\method{t}{NMsimModels}(x, ...)

\method{dimnames}{NMsimModels}(x, ...)

\method{rbind}{NMsimModels}(x, ...)

\method{cbind}{NMsimModels}(x, ...)
}
\arguments{
\item{x}{an NMsimModels object}

\item{...}{arguments passed to other methods.}
}
\value{
x stripped from the 'NMsimModels' class

logical if x is an 'NMsimModels' object

An object that is not of class 'NMsimModels'.
}
\description{
Remove NMsimModels class and discard NMsimModels meta data

Check if an object is 'NMsimModels'

Basic arithmetic on NMsimModels objects
}
\details{
When 'dimnames', 'merge', 'cbind', 'rbind', or 't' is
    called on an 'NMsimModels' object, the 'NMsimModels' class is dropped,
    and then the operation is performed. So if and 'NMsimModels' object
    inherits from 'data.frame' and no other classes (which is
    default), these operations will be performed using the
    'data.frame' methods. But for example, if you use 'as.fun' to
    get a 'data.table' or 'tbl', their respective methods are used
    instead.
}
