% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMsimModTab_utils.R
\name{unNMsimModTab}
\alias{unNMsimModTab}
\alias{is.NMsimModTab}
\alias{NMsimModTabOperations}
\alias{merge.NMsimModTab}
\alias{t.NMsimModTab}
\alias{dimnames.NMsimModTab}
\alias{rbind.NMsimModTab}
\alias{cbind.NMsimModTab}
\title{Remove NMsimModTab class and discard NMsimModTab meta data}
\usage{
unNMsimModTab(x)

is.NMsimModTab(x)

\method{merge}{NMsimModTab}(x, ...)

\method{t}{NMsimModTab}(x, ...)

\method{dimnames}{NMsimModTab}(x, ...)

\method{rbind}{NMsimModTab}(x, ...)

\method{cbind}{NMsimModTab}(x, ...)
}
\arguments{
\item{x}{an NMsimModTab object}

\item{...}{arguments passed to other methods.}
}
\value{
x stripped from the 'NMsimModTab' class

logical if x is an 'NMsimModTab' object

An object that is not of class 'NMsimModTab'.
}
\description{
Remove NMsimModTab class and discard NMsimModTab meta data

Check if an object is 'NMsimModTab'

Basic arithmetic on NMsimModTab objects
}
\details{
When 'dimnames', 'merge', 'cbind', 'rbind', or 't' is
    called on an 'NMsimModTab' object, the 'NMsimModTab' class is dropped,
    and then the operation is performed. So if and 'NMsimModTab' object
    inherits from 'data.frame' and no other classes (which is
    default), these operations will be performed using the
    'data.frame' methods. But for example, if you use 'as.fun' to
    get a 'data.table' or 'tbl', their respective methods are used
    instead.
}
