% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMsimRes_utils.R
\name{unNMsimRes}
\alias{unNMsimRes}
\alias{is.NMsimRes}
\alias{NMsimResOperations}
\alias{merge.NMsimRes}
\alias{t.NMsimRes}
\alias{dimnames.NMsimRes}
\alias{rbind.NMsimRes}
\alias{cbind.NMsimRes}
\title{Remove NMsimRes class and discard NMsimRes meta data}
\usage{
unNMsimRes(x)

is.NMsimRes(x)

\method{merge}{NMsimRes}(x, ...)

\method{t}{NMsimRes}(x, ...)

\method{dimnames}{NMsimRes}(x, ...)

\method{rbind}{NMsimRes}(x, ...)

\method{cbind}{NMsimRes}(x, ...)
}
\arguments{
\item{x}{an NMsimRes object}

\item{...}{arguments passed to other methods.}
}
\value{
x stripped from the 'NMsimRes' class

logical if x is an 'NMsimRes' object

An object that is not of class 'NMsimRes'.
}
\description{
Remove NMsimRes class and discard NMsimRes meta data

Check if an object is 'NMsimRes'

Basic arithmetic on NMsimRes objects
}
\details{
When 'dimnames', 'merge', 'cbind', 'rbind', or 't' is
    called on an 'NMsimRes' object, the 'NMsimRes' class is dropped,
    and then the operation is performed. So if and 'NMsimRes' object
    inherits from 'data.frame' and no other classes (which is
    default), these operations will be performed using the
    'data.frame' methods. But for example, if you use 'as.fun' to
    get a 'data.table' or 'tbl', their respective methods are used
    instead.
}
