% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMsimTab_utils.R
\name{NMsimTabOperations}
\alias{NMsimTabOperations}
\alias{merge.NMsimTab}
\alias{t.NMsimTab}
\alias{dimnames.NMsimTab}
\alias{rbind.NMsimTab}
\alias{cbind.NMsimTab}
\title{Basic arithmetic on NMsimTab objects}
\usage{
\method{merge}{NMsimTab}(x, ...)

\method{t}{NMsimTab}(x, ...)

\method{dimnames}{NMsimTab}(x, ...)

\method{rbind}{NMsimTab}(x, ...)

\method{cbind}{NMsimTab}(x, ...)
}
\arguments{
\item{x}{an NMsimTab object}

\item{...}{arguments passed to other methods.}
}
\value{
An object that is not of class 'NMsimTab'.
}
\description{
Basic arithmetic on NMsimTab objects
}
\details{
When 'dimnames', 'merge', 'cbind', 'rbind', or 't' is
    called on an 'NMsimTab' object, the 'NMsimTab' class is dropped,
    and then the operation is performed. So if and 'NMsimTab' object
    inherits from 'data.frame' and no other classes (which is
    default), these operations will be performed using the
    'data.frame' methods. But for example, if you use 'as.fun' to
    get a 'data.table' or 'tbl', their respective methods are used
    instead.
}
