% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMsimTestConf.R
\name{NMsimTestConf}
\alias{NMsimTestConf}
\title{Summarize and test NMsim configuration}
\usage{
NMsimTestConf(
  path.nonmem,
  dir.psn,
  method.execute,
  must.work = FALSE,
  system.type
)
}
\arguments{
\item{path.nonmem}{See ?NMsim}

\item{dir.psn}{See ?NMsim}

\item{method.execute}{See ?NMsim}

\item{must.work}{Throw an error if the configuration does not seem
to match system.}

\item{system.type}{See ?NMsim}
}
\description{
Summarize and test NMsim configuration
}
