% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMupdateFn.R
\name{NMupdateFn}
\alias{NMupdateFn}
\title{Update file names in control stream to match model name}
\usage{
NMupdateFn(
  x,
  section,
  model,
  fnext,
  add.section.text,
  par.file,
  text.section,
  quiet = FALSE
)
}
\arguments{
\item{x}{a control stream, path or `NMctl` object.}

\item{section}{What section to update}

\item{model}{Model name}

\item{fnext}{The file name extension of the file name to be
updated (e.g., one of "tab", "csv", "msf").}

\item{add.section.text}{Addditional text to insert right after
$SECTION. It can be additional TABLE variables.}

\item{par.file}{The Nonmem parameter that specifies the file. In
$TABLE, this is FILE. In $EST it's probably MSFO.}

\item{text.section}{This is used to overwrite the contents of the section. The section output file name will still handled/updated.}

\item{quiet}{Suppress messages? Default is `FALSE`.}
}
\description{
Update file names in control stream to match model name
}
\keyword{internal}
