% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMupdateInits.R
\name{NMupdateInits}
\alias{NMupdateInits}
\title{Create new Nonmem control stream with updated initial parameter values}
\usage{
NMupdateInits(file.mod, file.ext, newfile)
}
\arguments{
\item{file.mod}{The control stream to update. Will not be edited.}

\item{file.ext}{Path to ext file. Default is to replace extension on `file.mod`.}

\item{newfile}{New file to generate}
}
\value{
The resulting control stream path(s)
}
\description{
Create new Nonmem control stream with updated initial parameter values
}
\keyword{internal}
