% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMupdateSizes.R
\name{NMupdateSizes}
\alias{NMupdateSizes}
\title{Create or update $SIZES in a control stream}
\usage{
NMupdateSizes(file.mod, newfile = file.mod, wipe = FALSE, ...)
}
\description{
Create or update $SIZES in a control stream
}
\examples{
file.mod <- system.file("examples/nonmem/xgxr134.mod",package="NMsim")

NMupdateSizes(file.mod,newfile="~/newmod.mod",LTV=50)
}
