% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMwriteFilters.R
\name{NMwriteFilters}
\alias{NMwriteFilters}
\title{Write IGNORE/ACCEPT filters to NONMEM model}
\usage{
NMwriteFilters(file = NULL, lines = NULL, filters, write)
}
\arguments{
\item{file}{Path to control stream. Use `file` or `lines`.}

\item{lines}{Control stream as text lines. Use `file` or `lines`.}

\item{filters}{A data frome with filters, like returned by
`NMreadFilters()`.}

\item{write}{If `file` is provided, write the results to file? If
`lines` is used, `write` cannot be used.}
}
\value{
Resulting control stream (lines) as character vector
}
\description{
Write IGNORE/ACCEPT filters to NONMEM model
}
\keyword{internal}
