% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMwritePreamble.R
\name{NMwritePreamble}
\alias{NMwritePreamble}
\title{Update the pirana-style comments in top of a Nonmem control stream}
\usage{
NMwritePreamble(
  file.mod,
  description = NULL,
  based.on = NULL,
  author = NULL,
  write.file = TRUE
)
}
\arguments{
\item{file.mod}{The control stream to edit}

\item{description}{The desription to put in the preamble comments}

\item{based.on}{A control stream that the model was based on (will be a comment in preamble).}

\item{author}{Name of author to credit in preamble}

\item{write.file}{Write to file? If not, resulting control stream will be returned user as lines, and nothing else done.}
}
\value{
lines (charachter) for new control stream
}
\description{
Update the pirana-style comments in top of a Nonmem control stream
}
\keyword{@internal}
