% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padZeros.R
\name{padZeros}
\alias{padZeros}
\title{pad zeros on integers}
\usage{
padZeros(x, nchars)
}
\arguments{
\item{x}{integers to pad. They can be coded as characters already.}

\item{nchars}{Optional specification of length of character
strings to return. If not supplied, characters will be padded
to match length of max value in x.}
}
\value{
A character vector
}
\description{
pad zeros on integers
}
\keyword{internal}
