% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasteFuns.R
\name{pasteFuns}
\alias{pasteFuns}
\alias{pasteBegin}
\alias{pasteEnd}
\title{Paste string to start of vector only}
\usage{
pasteBegin(x, add, ...)

pasteEnd(x, add, ...)
}
\arguments{
\item{x}{A vector of strings}

\item{add}{A string to add}

\item{...}{Aditional arguments to `paste()`.}
}
\description{
paste(str,x) will prepend str to all values of x. use pasteBegin
to only paste it to the first value of x.
}
\keyword{internal}
