% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettyMatLines.R
\name{prettyMatLines}
\alias{prettyMatLines}
\title{Print OMEGA and SIGMA matrices for NONMEM sections in block format.
Note: This function currently only works with fixed blocks as in 
      the NMsim_NWPRI functionality for printing $THETAPV.}
\usage{
prettyMatLines(block_mat_string)
}
\arguments{
\item{block_mat_string}{Output of NMsim::NMcreateMatLines. This is a string of
OMEGA/SIGMA estimates that will be wrapped onto multiple lines for ease of
reading in NONMEM control streams.}
}
\value{
Character vector
}
\description{
Print OMEGA and SIGMA matrices for NONMEM sections in block format.
Note: This function currently only works with fixed blocks as in 
      the NMsim_NWPRI functionality for printing $THETAPV.
}
\keyword{internal}
