% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_NMsimRes.R
\name{print.summary_NMsimRes}
\alias{print.summary_NMsimRes}
\title{print method for NMsimRes summaries}
\usage{
\method{print}{summary_NMsimRes}(x, ...)
}
\arguments{
\item{x}{The summary object to be printed. See ?summary.NMsimRes}

\item{...}{Arguments passed to other print methods.}
}
\value{
NULL (invisibly)
}
\description{
print method for NMsimRes summaries
}
