% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prioritizePaths.R
\name{prioritizePaths}
\alias{prioritizePaths}
\title{first path that works}
\usage{
prioritizePaths(paths, must.work = FALSE)
}
\arguments{
\item{paths}{vector of file paths. Typically to Nonmem
executables.}

\item{must.work}{If TRUE, an error is thrown if no paths are
valid.}
}
\value{
A single path (character string)
}
\description{
When using scripts on different systems, the Nonmem path may
change from run to run. With this function you can specify a few
paths, and it will return the one that works on the system in use.
}
\examples{
library(NMdata)
NMdataConf(path.nonmem = prioritizePaths(c(
  "/opt/NONMEM/nm75/run/nmfe75",
  "C:/nm75g64/run/nmfe75.bat")
))

}
