% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMctl.R
\name{readCtl}
\alias{readCtl}
\title{Read as class NMctl}
\usage{
readCtl(x, ...)
}
\arguments{
\item{x}{object to read.}

\item{...}{Not used.}
}
\value{
An object of class `NMctl`.
}
\description{
Read as class NMctl
}
