% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleParsSimpar.R
\name{sampleParsSimpar}
\alias{sampleParsSimpar}
\title{Sample model parameters using the `simpar` package}
\usage{
sampleParsSimpar(file.mod, nsim, format = "ext")
}
\arguments{
\item{file.mod}{Path to model control stream. Will be used for
both `NMreadExt()` and `NMreadCov()`, and extension will
automatically be replaced by `.ext` and `.cov`.}

\item{nsim}{Number of sets of parameter values to generate. Passed
to `simpar`.}

\item{format}{"ext" (default) or "wide".}
}
\description{
Sample model parameters using the `simpar` package
}
