% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleCharArg.R
\name{simpleCharArg}
\alias{simpleCharArg}
\title{Check that a variable is a single character string meeting specified requirements}
\usage{
simpleCharArg(name.arg, val.arg, default, accepted, lower = TRUE, clean = TRUE)
}
\arguments{
\item{name.arg}{Name of the argument}

\item{val.arg}{argument value}

\item{default}{If val.arg is NULL, what should be returned?}

\item{accepted}{What values are allowed}

\item{lower}{run tolower?}

\item{clean}{clean white spaces?}
}
\value{
The resulting parameter value
}
\description{
Check that a variable is a single character string meeting specified requirements
}
\details{
Better options may be available in packages like
    checkmate. This function doesn't only check the parameter
    value, it also sets it to the default value if missing.
}
\keyword{internal}
