% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMsim_helpers.R
\name{simplePath}
\alias{simplePath}
\title{Simplify file paths by dropping .. and //}
\usage{
simplePath(path)
}
\arguments{
\item{path}{single or multiple file or dir paths as strings.}
}
\value{
Simplified paths as strings
}
\description{
Simplify file paths by dropping .. and //
}
\examples{
\dontrun{
path <- c("ds/asf.t","gege/../jjj.r")
NMsim:::simplePath(path)
}
}
\keyword{internal}
