% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPopEtas.R
\name{simPopEtas}
\alias{simPopEtas}
\title{Generate a population based on a Nonmem model}
\usage{
simPopEtas(
  file,
  N,
  seed.R,
  pars,
  file.phi,
  overwrite = FALSE,
  as.fun,
  file.mod,
  seed,
  ...
)
}
\arguments{
\item{file}{Passed to `NMdata::NMreadExt()`. Path to ext file. By
default, `NMreadExt()` uses a`auto.ext=TRUE` which means that
the file name extension is replaced by `.ext` (so you can normally specify a control stream path too). If your ext
file name extension is not `.ext`, add `auto.ext=FALSE` (see
...).}

\item{N}{Number of subjects to generate}

\item{seed.R}{Optional seed. Will be passed to `set.seed`. Same
thing as running `set.seed` just before calling
`simPopEtas()`.}

\item{pars}{A long-format parameter table containing par.type and
i columns. If this is supplied, the parameter values will not
be read from an ext file, and file has no effect. If an ext
file is available, it is most likely better to use the file
argument.}

\item{file.phi}{An optional phi file to write the generated
subjects to. This will generate a Nonmem-style `.phi` file that can be reused in combination with `NMsim_EBE`.}

\item{overwrite}{If `file.phi` exists already, overwrite it?
Default is `FALSE`.}

\item{as.fun}{The default is to return data as a data.frame. Pass
a function (say `tibble::as_tibble`) in as.fun to convert to
something else. If data.tables are wanted, use
as.fun="data.table". The default can be configured using
NMdataConf.}

\item{file.mod}{Deprecated. Use file instead.}

\item{seed}{Deprecated. Use seed.R instead.}

\item{...}{Additional arguments passed to NMdata::NMreadExt().}
}
\value{
A data.frame
}
\description{
Rather than simulating new subjects for each simulation, use this to create eta's that can be (re-)used for simulations with `NMsim_BEE()`. Although setting a seed for a simulation ensures reproducibility of a specific simulation, changing details of the data set is likely to break that reproducibility. Use `simPopEtas()` to ensure that a subject with a specific ID will always be simulated with the same eta values. This can also be used to reuse subjects between treatments in a simulation.
}
