% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringToSection.R
\name{stringToSection}
\alias{stringToSection}
\title{standardize section-denoting strings}
\usage{
stringToSection(x)
}
\arguments{
\item{x}{strings like THETA, $theta, $Theta etc (will all return THETA)}
}
\value{
character string
}
\description{
Always returns capital-case strings without $ or surrounding spaces
}
\keyword{internal}
