% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_NMsimRes.R
\name{summary.NMsimRes}
\alias{summary.NMsimRes}
\title{summary method for NMsim results (NMsimRes objects)}
\usage{
\method{summary}{NMsimRes}(object, ...)
}
\arguments{
\item{object}{An NMsimRes object (from NMsim).}

\item{...}{Not used}
}
\value{
A list with summary information on the NMsimRes object.
}
\description{
summary method for NMsim results (NMsimRes objects)
}
