% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMreadInits.R
\name{triagSize}
\alias{triagSize}
\title{Calculate number of elements for matrix specification}
\usage{
triagSize(diagSize)
}
\arguments{
\item{diagSize}{The length of the diagonal. Same as number of rows
or columns.}
}
\value{
An integer
}
\description{
calculate number of elements in the diagonal and lower triangle of
a squared matrix, based on the length of the diagonal.
}
\keyword{internal}
