% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typicalize.R
\name{typicalize}
\alias{typicalize}
\title{Set variability parameters to zero}
\usage{
typicalize(file.mod, lines, section, newfile)
}
\arguments{
\item{file.mod}{path to control stream to edit}

\item{lines}{control stream as lines. Use either file.sim or
lines.sim.}

\item{section}{The sections (parameter types) to edit. Default is
`c("OMEGA", "OMEGAP", "OMEGAPD")`.}

\item{newfile}{path and filename to new run. If missing or NULL,
output is returned as a character vector rather than written.}
}
\description{
Set variability parameters to zero
}
\keyword{internal}
