% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMsimTab_utils.R
\name{unNMsimTab}
\alias{unNMsimTab}
\title{Remove NMsimTab class and discard NMsimTab meta data}
\usage{
unNMsimTab(x)
}
\arguments{
\item{x}{An 'NMsimTab' object.}
}
\value{
x stripped from the 'NMsimTab' class
}
\description{
Remove NMsimTab class and discard NMsimTab meta data
}
