% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addEVID2.R
\name{addEVID2}
\alias{addEVID2}
\title{Add simulation records to dosing records}
\usage{
addEVID2(doses, time.sim, CMT)
}
\arguments{
\item{doses}{dosing records Nonmem style}

\item{time.sim}{A numerical vector with simulation times}

\item{CMT}{The compartment in which to insert the EVID=2 records}
}
\description{
Add simulation records to dosing records
}
