% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addEVID2.R
\name{addEVID2}
\alias{addEVID2}
\title{Add simulation (sample) records to dosing records (see `NMaddSamples()`)}
\usage{
addEVID2(...)
}
\arguments{
\item{...}{Passed to `NMaddSamples()`}
}
\value{
A data.frame
}
\description{
Add simulation (sample) records to dosing records (see `NMaddSamples()`)
}
