% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/needRun.R
\name{digestElements}
\alias{digestElements}
\title{Derive digests of argument values or their contents}
\usage{
digestElements(obj, funs)
}
\arguments{
\item{obj}{An object with elements to run digest on.}

\item{funs}{Named list of functions to be applied to elements (matched on names) in `obj`. Optional.}
}
\value{
A data.table with hashes/checksums.
}
\description{
Get digests of argument values. Optionally, functions can be run
on some arguments before calculating digests. An example would be
reading contents of a file where the file path is an argument.
}
