% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtapply.R
\name{dtapply}
\alias{dtapply}
\title{Apply function and return a data.table}
\usage{
dtapply(X, FUN, ...)
}
\arguments{
\item{...}{arguments passed to lapply}
}
\description{
A convenience function that returns a data.table with a column
representing the input values and a column with results. This is
still experimental and will not work for many input structures.
}
\examples{
NMsim:::dtapply(setNames(1:4,letters[1:4]),sqrt)
}
