% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findUpdated.R
\name{findUpdated}
\alias{findUpdated}
\title{Filter control streams to only those updated since last run}
\usage{
findUpdated(mods)
}
\arguments{
\item{mods}{list of (input or output) control streams to consider}
}
\value{
character vector of paths found models
}
\description{
Filter control streams to only those updated since last run
}
\keyword{internal}
