% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMreadInits.R
\name{initsToExt}
\alias{initsToExt}
\title{Convert inits elements to a parameter data.frame}
\usage{
initsToExt(elements)
}
\arguments{
\item{elements}{The elements object produced by `NMreadInits()`.}
}
\description{
Convert inits elements to a parameter data.frame
}
\keyword{internal}
