% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifyModel.R
\name{modifyModel}
\alias{modifyModel}
\title{Internal method for handling modify argument to NMsim}
\usage{
modifyModel(modify, dt.models = NULL, list.ctl = NULL)
}
\arguments{
\item{modify}{A list}

\item{dt.models}{a data.table}

\item{list.ctl}{List of coontrol streams as lines}
}
\value{
dt.models (data.table) or result list.ctl (list)
}
\description{
Internal method for handling modify argument to NMsim
}
\keyword{internal}
