% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMcreateDoses.R
\name{NMcreateDoses}
\alias{NMcreateDoses}
\title{Easily generate dosing records}
\usage{
NMcreateDoses(
  TIME,
  AMT = NULL,
  RATE = NULL,
  SS = NULL,
  CMT = 1,
  EVID = 1,
  addl = NULL
)
}
\arguments{
\item{TIME}{The time of the dosing events}

\item{AMT}{vector or dataa.frame with amounts amount}

\item{RATE}{Optional infusion rate}

\item{SS}{Optional steady-state flag}

\item{CMT}{Compartment number. Default is to dose into CMT=1.}

\item{EVID}{The event ID to use for doses. Default is to use
EVID=1, but EVID might also be wanted.}

\item{addl}{Optinal. A list of ADDL and II that will be applied to
last dose}
}
\description{
Combinations of different columns can be generated. Columns will
be extended by repeating last value of the column if needed in
order to match length of other columns.
}
\examples{
NMcreateDoses(TIME=c(0,1,4),AMT=c(2,1))
## Make Nonmem repeat the last dose. This is a total of 20 dosing events.
NMcreateDoses(TIME=c(0,12),AMT=c(2,1),addl=list(ADDL=9*2,II=12))

}
