% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMcreateMatLines.R
\name{NMcreateMatLines}
\alias{NMcreateMatLines}
\title{Create text lines for OMEGA and SIGMA Nonmem sections}
\usage{
NMcreateMatLines(omegas, type)
}
\arguments{
\item{omegas}{See NMreadExt and the pars element returned by that
function.}

\item{type}{The matrix type. OMEGA or SIGMA - case in-sensitive.}
}
\value{
Character vector
}
\description{
Create text lines for OMEGA and SIGMA Nonmem sections
}
\keyword{internal}
