% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMreadSim.R
\name{NMreadSim}
\alias{NMreadSim}
\title{Read simulation results based on NMsim's track of model runs}
\usage{
NMreadSim(x, as.fun)
}
\arguments{
\item{x}{Path to the simulation-specific rds file generated by
NMsim, typically called `NMsim_paths.rds`. Or a table of
simulation runs as returned by NMsim when `wait=FALSE`.}

\item{as.fun}{The default is to return data as a data.frame. Pass
a function (say tibble::as_tibble) in as.fun to convert to
something else. If data.tables are wanted, use
as.fun="data.table". The default can be configured using
NMdataConf.}
}
\value{
A data set of class defined by as.fun
}
\description{
Read simulation results based on NMsim's track of model runs
}
