% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMreadSimModTab.R
\name{NMreadSimModTab}
\alias{NMreadSimModTab}
\title{Read simulation results from rds objects and/or NMsimModTab objects}
\usage{
NMreadSimModTab(
  x,
  check.time = FALSE,
  dir.sims,
  wait = FALSE,
  quiet = FALSE,
  as.fun
)
}
\description{
Read simulation results from rds objects and/or NMsimModTab objects
}
\keyword{internal}
