% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMsim.R
\name{NMsim}
\alias{NMsim}
\title{Run simulations from an estimated Nonmem model}
\usage{
NMsim(
  path.mod,
  data,
  dir.sim,
  suffix.sim,
  order.columns = TRUE,
  script = NULL,
  subproblems,
  reuse.results = FALSE,
  seed,
  args.execute = "-clean=5",
  nmquiet = FALSE,
  text.table,
  type.input = "est",
  execute = TRUE,
  sge = FALSE,
  transform = NULL
)
}
\arguments{
\item{path.mod}{Path to the input control stream to run the
simulation on. The outpult control stream is for now assumed
to be stored next to the input control stream and ending in
.lst instead of .modl}

\item{data}{The simulation data as a data.frame.}

\item{dir.sim}{The directory in which NMsim will store all
generated files.}

\item{suffix.sim}{Give all filenames related to the simulation a
suffix. A short string describing the sim is recommended like
"ph3_regimens".}

\item{order.columns}{reorder columns by calling
NMdata::NMorderColumns before saving dataset and running
simulations? Default is TRUE.}

\item{script}{The path to the script where this is run.For
stamping of dataset so results can be traced back to code.}

\item{subproblems}{Number of subproblems to use as SUBPROBLEMS in
$SIMULATION block in Nonmem. The default is subproblem=0 which
means not to use SUBPROBLEMS.}

\item{reuse.results}{If simulation results found on file, should
they be used? If TRUE and reading the results fail, the
simulations will still be rerun.}

\item{seed}{Seed to pass to Nonmem}

\item{transform}{A list defining transformations to be applied
after the Nonmem simulations and before plotting. For each
list element, its name refers to the name of the column to
transform, the contents must be the function to apply. See
examples.}
}
\description{
Supply a data set and an input control stream, and NMsim will
create neccesary files, run the simulation and read the results.
}
