% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMsim.R
\name{NMsim}
\alias{NMsim}
\title{Run simulations from an estimated Nonmem model}
\usage{
NMsim(
  path.mod,
  data,
  dir.sim,
  suffix.sim,
  order.columns = TRUE,
  script = NULL,
  subproblems,
  reuse.results = FALSE,
  seed,
  args.execute = "-clean=5",
  nmquiet = FALSE,
  text.table,
  type.mod,
  execute = TRUE,
  sge = FALSE,
  transform = NULL,
  type.input
)
}
\arguments{
\item{path.mod}{Path to the input control stream to run the
simulation on. The outpult control stream is for now assumed
to be stored next to the input control stream and ending in
.lst instead of .modl}

\item{data}{The simulation data as a data.frame.}

\item{dir.sim}{The directory in which NMsim will store all
generated files.}

\item{suffix.sim}{Give all filenames related to the simulation a
suffix. A short string describing the sim is recommended like
"ph3_regimens".}

\item{order.columns}{reorder columns by calling
NMdata::NMorderColumns before saving dataset and running
simulations? Default is TRUE.}

\item{script}{The path to the script where this is run.For
stamping of dataset so results can be traced back to code.}

\item{subproblems}{Number of subproblems to use as SUBPROBLEMS in
$SIMULATION block in Nonmem. The default is subproblem=0 which
means not to use SUBPROBLEMS.}

\item{reuse.results}{If simulation results found on file, should
they be used? If TRUE and reading the results fail, the
simulations will still be rerun.}

\item{seed}{Seed to pass to Nonmem.}

\item{args.execute}{A charachter string that will be passed as
arguments PSN's `execute`.}

\item{nmquiet}{Silent messages from Nonmem.}

\item{text.table}{A character string including the variables to
export from Nonmem. The default is to export the same tables
as listed in the input control stream. But if many variables
are exported, and much fewer are used, it can speed up NMsim
significantly to only export what is needed (sometimes this is
as little as "PRED IPRED"). Nonmem writes data slowly so
reducing output data from say 100 columns to a handful makes a
big difference.}

\item{type.mod}{The control stream "type". Default is "est"
meaning that an $ESTIMATION block will be replaced by a
"$SIMULATION" block, and parameter estimates should be taken
from the estimation results. If the control stream has already
been turned into a simulation control stream, and only $INPUT,
$DATA, and $TABLE sections should be edited. This implies that
in case type.mod="sim", `subproblems` is
ignored. `type.mod` may be automated in the future.}

\item{execute}{Execute the simulation or only prepare it?
`execute=FALSE` can be useful if you want to do additional
tweaks or simulate using other parameter estimates.}

\item{sge}{Submit to cluster? Default is not to, but this is very
useful if creating a large number of simulations,
e.g. simulate with all parameter estimates from a bootstrap
result.}

\item{transform}{A list defining transformations to be applied
after the Nonmem simulations and before plotting. For each
list element, its name refers to the name of the column to
transform, the contents must be the function to apply.}

\item{type.input}{Deprecated. Use type.mod instead.}
}
\description{
Supply a data set and an input control stream, and NMsim will
create neccesary files, run the simulation and read the results.
}
