% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMupdateInits.R
\name{NMupdateInits}
\alias{NMupdateInits}
\title{Create new Nonmem control stream with updated initial parameter values}
\usage{
NMupdateInits(file.mod, newfile, fix)
}
\arguments{
\item{file.mod}{The control stream to update. Will not be edited.}

\item{newfile}{New file to generate}

\item{fix}{Fix the values? Probably only TRUE is supported.}
}
\value{
The resulting control stream contents as text
}
\description{
Create new Nonmem control stream with updated initial parameter values
}
\keyword{internal}
