% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMwriteInits.R
\name{NMwriteInits}
\alias{NMwriteInits}
\title{Writes a parameter table to a control stream}
\usage{
NMwriteInits(file.mod, update = TRUE, file.ext = NULL, values, newfile, ...)
}
\arguments{
\item{values}{A list of lists. Each list specifies a parameter
with named elements. par.type and i are required. j is
required for OMEGA and SIGMA. value, ll, ul and fix can be
supplied to modify the parameter. See examples.}
}
\description{
Writes a parameter table to a control stream
}
\examples{
values=list( list(par.type="THETA",i=2,value=1.4),
             list(par.type="THETA",i=3,FIX=1),
             list(par.type="OMEGA",i=2,j=2,value=0.1))
}
