% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifyHelpers.R
\name{overwrite}
\alias{overwrite}
\title{Create function that modifies text elements in a vector}
\usage{
overwrite(...)
}
\arguments{
\item{...}{Passed to `gsub()`}
}
\value{
A function that runs `gsub` to character vectors
}
\description{
Create function that modifies text elements in a vector
}
\examples{
myfun <- overwrite("b","d")
myfun(c("a","b","c","abc"))
## regular expressions
myfun2 <- overwrite("b.*","d")
myfun2(c("a","b","c","abc"))
}
