% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleParsSimpar.R
\name{sampleParsSimpar}
\alias{sampleParsSimpar}
\title{Sample model parameters using the `simpar` package}
\usage{
sampleParsSimpar(file.mod, nsim, format = "ext", seed.R, as.fun)
}
\arguments{
\item{file.mod}{Path to model control stream. Will be used for
both `NMreadExt()` and `NMreadCov()`, and extension will
automatically be replaced by `.ext` and `.cov`.}

\item{nsim}{Number of sets of parameter values to generate. Passed
to `simpar`.}

\item{format}{"ext" (default) or "wide".}

\item{seed.R}{seed value passed to set.seed().}

\item{as.fun}{The default is to return data as a data.frame. Pass
a function (say `tibble::as_tibble`) in as.fun to convert to
something else. If data.tables are wanted, use
as.fun="data.table". The default can be configured using
NMdataConf.}
}
\value{
A table with sampled model parameters
}
\description{
Sample model parameters using the `simpar` package
}
\author{
Sanaya Shroff, Philip Delff
}
