% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPopEtas.R
\name{simPopEtas}
\alias{simPopEtas}
\title{Generate a population based on a Nonmem model}
\usage{
simPopEtas(file.mod, N, seed, file.phi, as.fun)
}
\arguments{
\item{file.mod}{Path to input control stream}

\item{N}{Number of subjects to generate}

\item{seed}{Optional seed. Will be passed to `set.seed`. Same
thing as running `set.seed` just before calling
`simPopEtas()`.}

\item{file.phi}{An optional phi file to write the generated subjects to.}

\item{as.fun}{The default is to return data as a data.frame. Pass
a function (say `tibble::as_tibble`) in as.fun to convert to
something else. If data.tables are wanted, use
as.fun="data.table". The default can be configured using
NMdataConf.}
}
\description{
Generate a population based on a Nonmem model
}
