% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeCovs.R
\name{summarizeCovs}
\alias{summarizeCovs}
\title{Summarize simulated exposures relative to reference subject (see `forestSummarize()`)}
\usage{
summarizeCovs(...)
}
\arguments{
\item{...}{Passed to `forestSummarize()`}
}
\value{
A data.frame
}
\description{
Summarize simulated exposures relative to reference subject (see `forestSummarize()`)
}
